Reorganized the template files
